New storage load playbooks
This playbooks create and verify storage load on a deployed setup
The idea is to add storage load in any upgrade job,
usage:
call the "create" after the 1st deployment - for example in a dedicated job stage.
call the "verify" after the upgrade
supported tags are [swift,cinder,glance] , default is all.

Change-Id: Ia2dbbf9c0bc16df78432a818c5c7cdc977ba3b31